Add kerberos into to p10k prompt.
Add a function, prompt_my_krb_validity, that inject the current
validity, in days:hours:minutes, and the extensible lifetime of the
current kerberos ticket, using the same format.
Enabled only on CERN machines.
The prompt has to be activated by adding my_krb_validity to the list of
prompt contained in the variable POWERLEVEL9K_RIGHT_PROMPT_ELEMENTS.
To install it permanently, edit the file .p10k.zsh and apply the changes
there.